requirements.txt
flask
flask-sqlalchemy
flask-cors
psycopg2-binary
pydantic
pydantic-settings
sqlalchemy
langchain
langchain-groq
langchain-huggingface
langchain-community
langchain-chroma
sentence-transformers
pdfplumber
PyPDF2
pandas
numpy
python-dotenv
uvicorn
gunicorn
faker
chromadb
faiss-cpu
greenlet==3.0.0